requirements.txt
avro-python3==1.8.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
confluent-kafka
idna==2.10
influxdb==5.3.0
pandas
paramiko==2.7.1
psycopg2==2.8.5
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
sshtunnel==0.1.5
urllib3==1.25.10
